Removal of several functions and methods. The source code folder with the sln file has been added back to GitHub for portability for people who wish to see and/or edit the code.
Removed several functions which were not necessary or were redundant (eg nil and empty for the Maybe type).
Removed some functions that are just existing functions with default/specialized implementations.
Removed several methods from ComputationBuilders (eg TryWith, TryFinally, Run, Delay). The builder classes are not sealed so, if needed, inheritance can be used to get the desired additional behavior. What was removed was not needed for a simple monadic workflow.
Made some functions/methods non-inlined that were before where the inlining was unnecessary.
Renamed the Proxy type to Tagged and added a few utility functions to it.